adding the starts of unflow on edit
However, spotted that the current flow algorithm (or rather
the reflow) is not 100% robust. For edge cases with chains
longer than two nodes, the flow might not work correctly.

Will stop work on the unflow and start adding unit tests to
prove every chained case, so we can be sure even on weird
documents things will flow correctly